Combine RANGE with CHAR
`[a-z_]` is a valid expression, meaning an `a-z` range or a `_`.

We can simplify the rule by making a single `CHAR` rule whose value
is an array, containing either a single code point matching a character,
or an array of two numbers denoting a range of code points.

This means that the only rules returned are `CHAR` and `END`, which can probably
be further simplified into a simple numeric array return, with a special identifier
(`0`? `null`?) denoting an END token.